Move utils and persistentjson to common.
