+++ lib/atom_listener.js
@@ -7,11 +7,12 @@ const atomUtils = require("./atom_utils");
 const bufferAction = require("./common/buffer_action");
 const CompositeDisposable = require("atom").CompositeDisposable;
 const editorAction = require("./common/editor_action");
+const messageAction = require("./common/message_action");
 const floop = require("./common/floop");
 const fs = require("fs-plus");
 const path = require("path");
 const prefs = require("./common/userPref_model");
-const utils = require("./utils");
+const utils = require("./common/utils");
 const ignore = require("./common/ignore");
 
 const IMAGINARY_ID = -1;
