+++ lib/common/persistentjson.js
