+++ lib/floobits.js
@@ -13,7 +13,7 @@ const _ = require("lodash");
 const api = require("./common/api");
 const atomUtils = require("./atom_utils");
 const message_action = require("./common/message_action");
-const utils = require("./utils");
+const utils = require("./common/utils");
 const usersModel = require("./common/user_model");
 const prefs = require("./common/userPref_model");
 const FlooUrl = require("./floourl").FlooUrl;
@@ -592,7 +592,7 @@ module.exports = {
     fl.base_path = floobitsPath;
     fl.floourl = floourl;
 
-    const PersistentJson = require("./persistentjson");
+    const PersistentJson = require("./common/persistentjson");
     const persistentJson = new PersistentJson();
     persistentJson.load();
     persistentJson.update(floobitsPath, url);
