+++ lib/follow_view.js
@@ -6,7 +6,7 @@ const $$ = require("atom-space-pen-views").$$;
 const editorAction = require("./common/editor_action");
 const prefs = require("./common/userPref_model");
 const SelectListView = require("./select_list_view");
-const utils = require("./utils");
+const utils = require("./common/utils");
 const EVERYONE = "all changes";
 
 var FollowView = function (me, users) {
