+++ lib/media_sources.js
@@ -3,7 +3,7 @@
 const $$ = require("atom-space-pen-views").$$;
 const _ = require("lodash");
 
-const utils = require("./utils");
+const utils = require("./common/utils");
 const SelectListView = require("./select_list_view");
 
 function MediaSources (sources, cb) {
