+++ lib/recentworkspaceview.js
@@ -3,8 +3,8 @@
 const $$ = require("atom-space-pen-views").$$;
 const _ = require("lodash");
 
-const utils = require("./utils");
-const PersistentJson = require("./persistentjson");
+const utils = require("./common/utils");
+const PersistentJson = require("./common/persistentjson");
 const SelectListView = require("./select_list_view");
 
 function RecentWorkspaceView () {
