+++ lib/select_directory_to_share.js
@@ -2,7 +2,7 @@
 
 const $$ = require("atom-space-pen-views").$$;
 
-const utils = require("./utils");
+const utils = require("./common/utils");
 const SelectListView = require("./select_list_view");
 
 function DirectorySelectorView (directories, cb) {
