+++ lib/select_list_view.js
@@ -3,7 +3,7 @@
 
 const AtomSelectListView = require("atom-space-pen-views").SelectListView;
 
-const utils = require("./utils");
+const utils = require("./common/utils");
 
 
 function SelectListView (cb) {
